Login & Registration (LARGE COMMIT)
I have done a lot in terms of Python Conversion, hence me taking 9 hours give or take a few quarters. I still need to figure out how to store user login states, but that's something for another day. Your issue on the login and registration pages is still valid since this commit does not fix/complete that issue. After I am done with login and registration pages, I will be able to easily do the upload page.  For user login state, I have variables named loggedIn and currentuser for customized messages such as Hello, fawkes-c if you are logged in as fawkes-c. But this only works for one user. We need a database for user login states or use the flask-login module which I can't figure out how to implement and probably will tomorrow. The upload page requires user login states so I need to do that first.
Have a good day